refactor - move hevm utils to its own file, add workflow job
