+++ setup.cfg
@@ -26,6 +26,7 @@ line_length = 100
 [tool:pytest]
 addopts = -n auto
 	--dist worksteal
+  --strict-markers
 python_files = test_*.py
 testpaths = tests
 xfail_strict = true
@@ -33,7 +34,7 @@ markers =
 	fuzzing: Run Hypothesis fuzz test suite (deselect with '-m "not fuzzing"')
 	requires_evm_version(version): Mark tests that require at least a specific EVM version and would throw `EvmVersionException` otherwise
 	venom_xfail: mark a test case as a regression (expected to fail) under the venom pipeline
-
+	hevm: run tests marked for symbolic execution
 
 [coverage:run]
 branch = True
