+++ tests/unit/compiler/venom/test_algebraic_binopt.py
@@ -1,17 +1,15 @@
-import subprocess
-
 import pytest
 
+from tests.hevm import hevm_check
 from tests.venom_utils import assert_ctx_eq, parse_from_basic_block
-from vyper.ir.compile_ir import assembly_to_evm
-from vyper.venom import StoreExpansionPass, VenomCompiler
-from vyper.venom.basicblock import IRLiteral, IRInstruction
 from vyper.venom.analysis import IRAnalysesCache
 from vyper.venom.passes import AlgebraicOptimizationPass, StoreElimination
 
 """
 Test abstract binop+unop optimizations in algebraic optimizations pass
 """
+
+pytestmark = pytest.mark.hevm
 
 
 def _sccp_algebraic_runner(pre, post, hevm=True):
@@ -27,66 +25,6 @@ def _sccp_algebraic_runner(pre, post, hevm=True):
 
     if hevm:
         hevm_check(pre, post)
-
-
-def _prep_hevm(venom_source_code):
-    ctx = parse_from_basic_block(venom_source_code)
-
-    num_calldataloads = 0
-    num_return_values = 0
-    for fn in ctx.functions.values():
-        for bb in fn.get_basic_blocks():
-            for inst in bb.instructions:
-                # transform `param` instructions into "symbolic" values for
-                # hevm via calldataload
-                if inst.opcode == "param":
-                    # hevm limit: 256 bytes of symbolic calldata
-                    assert num_calldataloads < 8
-
-                    inst.opcode = "calldataload"
-                    inst.operands = [IRLiteral(num_calldataloads * 32)]
-                    num_calldataloads += 1
-
-            term = bb.instructions[-1]
-            # test convention, terminate by `return`ing the variables
-            # you want to check
-            assert term.opcode == "return"
-            num_return_values = 0
-            for op in term.operands:
-                ptr = IRLiteral(num_return_values * 32)
-                new_inst = IRInstruction("mstore", [op, ptr])
-                bb.insert_instruction(new_inst, index=-1)
-                num_return_values += 1
-
-            # return 0, 32 * num_variables
-            term.operands = [IRLiteral(num_return_values * 32), IRLiteral(0)]
-
-        ac = IRAnalysesCache(fn)
-        # requirement for venom_to_assembly
-        StoreExpansionPass(ac, fn).run_pass()
-
-    compiler = VenomCompiler([ctx])
-    return assembly_to_evm(compiler.generate_evm(no_optimize=True))[0].hex()
-
-def hevm_check(pre, post):
-    # perform hevm equivalence check
-    print("HEVM COMPARE.")
-    print("BEFORE:", pre)
-    print("OPTIMIZED:", post)
-    bytecode1 = _prep_hevm(pre)
-    bytecode2 = _prep_hevm(post)
-
-    # debug:
-    print("RUN HEVM:")
-    print(bytecode1)
-    print(bytecode2)
-
-    subp_args = ["hevm", "equivalence", "--code-a", bytecode1, "--code-b", bytecode2]
-    # quiet:
-    # subprocess.check_output(["hevm", "equivalence", "--code-a", bytecode1, "--code-b", bytecode2])
-
-    # verbose:
-    subprocess.check_call(subp_args)
 
 
 def test_sccp_algebraic_opt_sub_xor():
